add index.blade

+++ database/migrations/2024_03_25_154544_create_comics_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("title", 50);
             $table->text("description");
             $table->string("thumb", 100);
-            $table->tinyInteger("price");
+            $table->float("price");
             $table->string("series", 50);
             $table->dateTime("sale_date");
             $table->string("type", 30);
